Switched from templates to views
The main issue was that it does not seem to be possible to use
send_from_directory with the template_folder. Since I am trying to
implement some form of MVC, the name views made sense. Placing it in
static was expected to be needed, because send_form_directory seems to
work on that directory.