Fixed the issue that it either worked standalone or through mod_wsgi
The real issue was that to make it pep8 compliant, it is needed to have
all the imports at the top, to resolve the difference in paths, the
module directory needs to be added to the path and afterwards the
modules can be loaded from the various packages. To 'resolve' the issue,
dynamic loading is used, and now it works when starting standalone as
well as when running the application through mod_wsgi, while still being
pep8 compliant.