Cleaned controllers
Made the controllers more consistent and clean. Did some updates to the
addFoodItem methods, tried to make it possible to assign a string to a
property and implicitly fill the relation, still not quite sure how to
do this, the problem with this approach seems to be that for both
getting and setting a lot of roundtrips to the database are  needed,
which does not scale.