Added libraries for ngMaterial
Angular material seems to be the preferred way of making the application
more beatiful. Since it seems that not all libraries are hosted by
google anymore, the locations are adjusted to code.angularjs.org. It
might be worthwhile to consider the use of an alternative framework, but
further investigation is needed, especially to determine if it should be
the latest version of Angualr, or an alternative altogether.